ExceptionThrower: Add a limit the number of times an exception is thrown
Constantly throwing the same exception every time a method is called
simulates a complete failure for the app. That's a useful failure mode
to test, but an equally interesting one is whether the app is able to
recover from intermittent problems. And for that, we need the ability to
limit the number of times an exception is thrown for a given method.